refactor(features): move out `tweets` folder out of `home`
As this feature can be used not only on home page
