+++ src/features/tweets/ui/LikeTweetButton/interface.ts
